+++ Library/Homebrew/os/mac/pathname.rb
@@ -1,13 +1,5 @@
-require "os/mac/shared_mach"
+require "os/mac/mach"
 
 class Pathname
-  if !ENV["HOMEBREW_NO_RUBY_MACHO"]
-    require "os/mac/ruby_mach"
-    include RubyMachO
-  else
-    require "os/mac/cctools_mach"
-    include CctoolsMachO
-  end
-
-  include SharedMachO
+  include MachO
 end
